[NUI] Remove VectorAnimationView, Add LottieAnimationView
